Adjusted permissions for products and reviews calls, implemented review by product call

+++ src/main/java/it/unibz/infosec/examproject/product/domain/ManageProducts.java
@@ -46,8 +46,11 @@ public class ManageProducts {
         return productRepository.save(product);
     }
 
-    public Product deleteProduct(Long id) {
+    public Product deleteProduct(Long id, Long vendorId) {
         final Product product = validateProduct(id);
+        if (!vendorId.equals(product.getVendorId())) {
+            throw new IllegalArgumentException("Only the owner of a product can delete it");
+        }
         productRepository.delete(product);
         return product;
     }
